Updated CreditExpert details
CreditExpert FAQ is on a different domain to the main site which means that the justdelete.me icon doesn’t appear in the address bar. I have changed the URL and moved the cancellation link into the notes — there’s nothing in there that isn’t already mentioned but I guess its useful to have.

Feature request: It would be good to have separate Site & Cancellation URLs in the json feed.