Add entry for mailbox.org to sites.json
No translations, as as of just having it tested right now in several languages, the "Cancel Account" menu is not there in other languages than German and English, or I just could't see it! And their site seems to have some issues in general when it comes to the translations as some menus and submenus are not translated even after logging out-and-in and read in English.